requirements.txt
# Core web framework
fastapi>=0.110
uvicorn[standard]>=0.27
websockets>=12.0

# Desktop GUI
pywebview>=5.0

# Scraping and downloading
selenium>=4.15
undetected-chromedriver>=3.5
requests>=2.31

# PDF processing
PyMuPDF>=1.23
pytesseract>=0.3.10
Pillow>=10.0

# Search and indexing
whoosh>=2.7.4

# LLM integration (uses Ollama HTTP API directly, no torch needed)
numpy>=1.24

# Data models and utilities
pydantic>=2.5
rich>=13.0
tqdm>=4.66
